+++ shiny/templates/chat/llm-enterprise/aws-bedrock-anthropic/app.py
@@ -28,6 +28,9 @@ ui.page_opts(
 chat = ui.Chat(id="chat")
 chat.ui()
 
+# Store chat state in the url when an "assistant" response occurs
+chat.enable_bookmarking(chat_client, bookmark_store="url")
+
 
 # Define a callback to run when the user submits a message
 @chat.on_user_submit