+++ shiny/templates/chat/llms/google/app.py
@@ -29,6 +29,9 @@ ui.page_opts(
 chat = ui.Chat(id="chat")
 chat.ui()
 
+# Store chat state in the url when an "assistant" response occurs
+chat.enable_bookmarking(chat_client, bookmark_store="url")
+
 
 # Generate a response when the user submits a message
 @chat.on_user_submit